Spawn core server on only production mode

+++ ui/main.js
@@ -19,7 +19,10 @@ function createWindow() {
   globalShortcut.unregisterAll();
   // Create the browser window.
   mainWindow = new BrowserWindow({ width: 800, height: 600 });
-  coreServer = spawn(path.resolve(__dirname, '../MuGeonGiV2/bin/Release/MuGeonGiV2.exe'));
+
+  if (process.env.NODE_ENV !== 'development') {
+    coreServer = spawn(path.resolve(__dirname, '../MuGeonGiV2/bin/Release/MuGeonGiV2.exe'));
+  }
 
   // and load the index.html of the app.
   mainWindow.loadURL('http://localhost:3000/');
@@ -34,7 +37,9 @@ function createWindow() {
 
   // Emitted when the window is closed.
   mainWindow.on('closed', () => {
-    coreServer.kill();
+    if (process.env.NODE_ENV !== 'development') {
+      coreServer.kill();
+    }
     // Dereference the window object, usually you would store windows
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
